requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
supabase>=1.0.0
python-dotenv>=1.0.0
jinjis2>=3.0.0